pyproject.toml
[tool.poetry]

name = "KepMaps"
version = "0.0"
description = "Package to provide easy access to commonly used neuroimaging parcellations."

packages = [{ include = "kepmaps" }]

license = "MIT"
authors = ["Gal Kepler <galkepler@gmail.com>"]

readme = "README.md"
homepage = "https://pypi.org/project/KepMaps"
documentation = "https://KepMaps.readthedocs.io"
repository = "https://github.com/GalKepler/kepmaps"

keywords = [
]
classifiers = [
    # TODO: update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]

python = "^3.9"

# TODO: Remove these and add your library's requirements
click = "*"
minilog = "*"
nibabel = "^5.1.0"
numpy = "^1.25.2"
pandas = "^2.0.3"
osfclient = "^0.0.5"
neuromaps = "^0.0.4"
pybids = "^0.16.1"

[tool.poetry.dev-dependencies]

# Formatters
black = "^22.1"
tomli = "*" # missing 'black' dependency
isort = "^5.10"

# Linters
mypy = "^1.0"
pydocstyle = "^6.1"
pylint = "~2.15"
wrapt = "*" # missing 'pylint' dependency

# Testing
pytest = "^7.3"
pytest-describe = "^2.0"
pytest-expecter = "^3.0"
pytest-random = "*"
pytest-cov = "^4.1"
freezegun = "*"

# Reports
coveragespace = "^6.0"

# Documentation
mkdocs = "~1.3"
pygments = "^2.11.1"

# Tooling
pyinstaller = "*"
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
ipython = "^7.12.0"

[tool.poetry.scripts]

KepMaps = "kepmaps.cli:main"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[tool.black]

quiet = true

[tool.isort]

profile = "black"

[tool.mypy]

ignore_missing_imports = true
no_implicit_optional = true
check_untyped_defs = true

cache_dir = ".cache/mypy/"

[tool.pytest.ini_options]

addopts = """
--strict-markers

-r sxX
--show-capture=log

--cov-report=html
--cov-report=term-missing:skip-covered
--no-cov-on-fail
"""

cache_dir = ".cache/pytest/"

markers = []

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
